requirements.txt
coinbase-advanced-py
python-dotenv
prefect